backend/requirements.txt
ccxt==4.0.0
pandas==2.0.3
numpy==1.24.3
google-cloud-bigquery==3.11.4
python-dotenv==1.0.0
scipy==1.10.1
newspaper3k==0.2.8
nltk==3.8.1
ta==0.10.2
matplotlib==3.7.2
requests==2.31.0